spike: represent a directory with a trailing /
The artifact map in Holos has no model for a directory.  Until now
transformers have always produced one single file as an output.

For the kubectl-slice use case we need to model a directory as an output
artifact.

This is a quick spike to explore the use case and work with spiarh on
developing it further.  It technically works so long as the slice
transformer is the final one and there are no validators.  If it's good
then we'll need to define the semantics for passing an entire directory
from one transformer to the next.

The trailing slash hack may actually be fine, but I need to think about
it and discuss it further before committing it to the core schema.